add new stragegy

+++ config/arbitrage.go
@@ -3,5 +3,5 @@ package config
 type Arbitrage struct {
 	RpcURL     string `mapstructure:"rpc_url"`
 	PrivateKey string `mapstructure:"private_key"`
-	ChainID    int    `mapstructure:"chain_id"`
+	ChainID    int64  `mapstructure:"chain_id"`
 }
